renamed main to bootstrap and moved out logging.
Rationale: main is actionally in err.py and was colliding. This module
is more about the bootstrap of Errbot than a main.

logging was cluttering err.py
