+++ errbot/backends/test.py
@@ -13,7 +13,7 @@ from errbot.rendering import text
 from errbot.backends.base import Message, Room, Person, RoomOccupant, ONLINE
 from errbot.core_plugins.wsview import reset_app
 from errbot.core import ErrBot
-from errbot.main import setup_bot
+from errbot.bootstrap import setup_bot
 
 # Can't use __name__ because of Yapsy
 log = logging.getLogger('errbot.backends.test')
