+++ errbot/bootstrap.py
@@ -139,7 +139,15 @@ def enumerate_backends(config):
     return [plug.name for (_, _, plug) in bpm.getPluginCandidates()]
 
 
-def main(bot_class, logger, config, restore=None):
+def bootstrap(bot_class, logger, config, restore=None):
+    """
+    Main starting point of Errbot.
+
+    :param bot_class: The backend class inheriting from Errbot you want to start.
+    :param logger: The logger you want to use.
+    :param config: The config.py module.
+    :param restore: Start Errbot in restore mode (from a backup).
+    """
     bot = setup_bot(bot_class, logger, config, restore)
     log.debug('Start serving commands from the %s backend' % bot.mode)
     bot.serve_forever()
