+++ errbot/logging.py
@@ -1,0 +1,40 @@
+import inspect
+import logging
+
+
+def ispydevd():
+    for frame in inspect.stack():
+        if frame[1].endswith("pydevd.py"):
+            return True
+    return False
+
+
+root_logger = logging.getLogger()
+logging.getLogger('yapsy').setLevel(logging.INFO)  # this one is way too verbose in debug
+logging.getLogger('Rocket.Errors.ThreadPool').setLevel(logging.INFO)  # this one is way too verbose in debug
+root_logger.setLevel(logging.INFO)
+
+pydev = ispydevd()
+stream = sys.stdout if pydev else sys.stderr
+isatty = pydev or stream.isatty()  # force isatty if we are under pydev because it supports coloring anyway.
+console_hdlr = logging.StreamHandler(stream)
+
+if isatty:
+    from colorlog import ColoredFormatter  # noqa
+    formatter = ColoredFormatter(
+        "%(asctime)s %(log_color)s%(levelname)-8s%(reset)s "
+        "%(blue)s%(name)-25.25s%(reset)s %(white)s%(message)s%(reset)s",
+        datefmt="%H:%M:%S",
+        reset=True,
+        log_colors={
+            'DEBUG': 'cyan',
+            'INFO': 'green',
+            'WARNING': 'yellow',
+            'ERROR': 'red',
+            'CRITICAL': 'red',
+        }
+    )
+    console_hdlr.setFormatter(formatter)
+else:
+    console_hdlr.setFormatter(logging.Formatter("%(asctime)s %(levelname)-8s %(name)-25s %(message)s"))
+root_logger.addHandler(console_hdlr)
