+++ tests/backend_manager_tests.py
@@ -3,7 +3,7 @@ import unittest
 import logging
 
 from errbot.core import ErrBot
-from errbot.main import CORE_BACKENDS
+from errbot.bootstrap import CORE_BACKENDS
 from errbot.specific_plugin_manager import SpecificPluginManager
 
 logging.basicConfig(level=logging.DEBUG)
