+++ tests/utils_tests.py
@@ -4,7 +4,7 @@ import pytest
 
 from errbot.backends.test import ShallowConfig
 from errbot.core import bot_config_defaults
-from errbot.main import CORE_STORAGE
+from errbot.bootstrap import CORE_STORAGE
 from errbot.specific_plugin_manager import SpecificPluginManager
 from errbot.storage.base import StoragePluginBase
 from errbot.utils import *
